sdk/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "snapper-sdk"
version = "1.0.0"
description = "Python SDK for Snapper Agent Application Firewall"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "httpx>=0.26.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.20.0"]
gemini = ["google-generativeai>=0.4.0"]
all = [
    "openai>=1.0.0",
    "anthropic>=0.20.0",
    "google-generativeai>=0.4.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.urls]
Homepage = "https://github.com/jmckinley/snapper"
Documentation = "https://github.com/jmckinley/snapper/blob/main/docs/AI_PROVIDERS.md"

[tool.setuptools.packages.find]
include = ["snapper*"]
